+++ ChangeLog.md
@@ -3,6 +3,8 @@ unreleased changes
 * Fix bug, introduced in v1.4.0, where attachment names were getting mangled
 * Correctly set default MCI for ServerTemplates on download
 * Improve formatting of description fields when downloading ServerTemplates
+* Default empty inputs to to "inherit from RightScript" rather than be a blank
+  value when downloading ServerTemplates
 
 v1.4.3 / 2016-08-11
 -------------------